+++ service/src/main/java/com/microservice/service/controller/ServiceController.java
@@ -3,6 +3,8 @@ package com.microservice.service.controller;
 import com.microservice.service.entity.Service;
 import com.microservice.service.services.IService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,26 +17,39 @@ public class ServiceController {
     private IService service;
 
     @GetMapping
-    public List<Service> findAllServices(){
-        return service.findAll();
+    public ResponseEntity<?> findAllServices(){
+        List<Service> services = service.findAll();
+        if(services.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not services yet");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(services);
     }
     @GetMapping("/{id}")
-    public Service findServiceById(@PathVariable("id") Integer id){
-        return service.findById(id);
+    public ResponseEntity<?> findServiceById(@PathVariable("id") Integer id){
+        Service serviceFound = service.findById(id);
+        if(serviceFound == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Service not found");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(serviceFound);
     }
 
     @PostMapping
-    public Service createService(Service srv){
-        return service.create(srv);
+    public ResponseEntity<?> createService(Service srv){
+        return ResponseEntity.status(HttpStatus.CREATED).body(service.create(srv));
     }
 
     @PutMapping
-    public Service editService(Service srv){
-        return service.update(srv);
+    public ResponseEntity<?> editService(Service srv){
+        return ResponseEntity.status(HttpStatus.CREATED).body(service.update(srv));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteService(@PathVariable("id") Integer id){
+    public ResponseEntity<?> deleteService(@PathVariable("id") Integer id){
+        Service serviceFound = service.findById(id);
+        if(serviceFound == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Service not found");
+        }
         service.remove(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Deleted");
     }
 }
