+++ service/src/main/java/com/microservice/service/controller/SpecialtyController.java
@@ -3,6 +3,8 @@ package com.microservice.service.controller;
 import com.microservice.service.entity.Specialty;
 import com.microservice.service.services.SpecialtyService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,30 +16,40 @@ public class SpecialtyController {
     private SpecialtyService service;
 
     @GetMapping
-    public List<Specialty> findAllSpecialties(){
-        return service.findAll();
+    public ResponseEntity<?> findAllSpecialties(){
+        List<Specialty> specialties = service.findAll();
+        if(specialties.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not doctors yet");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(specialties);
     }
 
     @GetMapping("/{id}")
-    public Specialty findSpecialtyById(@PathVariable("id") Integer id){
-        return service.findById(id);
+    public ResponseEntity<?> findSpecialtyById(@PathVariable("id") Integer id){
+        Specialty specialtyFound = service.findById(id);
+        if(specialtyFound == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Doctor not found");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(specialtyFound);
     }
 
     @PostMapping
-    public Specialty createSpecialty(Specialty specialty){
-        return service.create(specialty);
+    public ResponseEntity<?> createSpecialty(Specialty specialty){
+        return ResponseEntity.status(HttpStatus.CREATED).body(service.save(specialty));
     }
 
     @PutMapping
-    public Specialty editSpecialty(Specialty specialty){
-        return service.edit(specialty);
+    public ResponseEntity<?> editSpecialty(Specialty specialty){
+        return ResponseEntity.status(HttpStatus.ACCEPTED).body(service.edit(specialty));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteSpecialty(@PathVariable("id") Integer id){
+    public ResponseEntity<?> deleteSpecialty(@PathVariable("id") Integer id){
+        Specialty specialtyFound = service.findById(id);
+        if(specialtyFound == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Specialty not found");
+        }
         service.remove(id);
+        return ResponseEntity.noContent().build();
     }
-
-
-
 }
